+++ main.lua
@@ -18,7 +18,7 @@ import "android.graphics.Typeface"
 import "res"
 
 --以下是封装好的各种各样的代码。如需使用请查看 README.md 以及 layout.aly 最后的demo
-local isAndroidX=false
+isAndroidX=false
 xpcall(function()
   import "androidx.appcompat.widget.Toolbar"
   import "androidx.appcompat.view.ContextThemeWrapper"
@@ -28,6 +28,19 @@ end,
 function()
   activity.setTheme(android.R.style.Theme_Material_Light)
 end)
+
+pcall(function()
+  androidhwext={R=luajava.bindClass("androidhwext.R")}
+  isEmui=true
+end)
+
+--EMUI系统会有额外的 R 类
+if isEmui then
+  androidhwext.res=res.getOrNewResWithRClass(androidhwext.R)
+end
+
+local SDK_INT=Build.VERSION.SDK_INT
+isSdk31AndUp=SDK_INT>=31
 
 local codeTypeface=Typeface.createFromFile(activity.getLuaDir().."/JetBrainsMono-Regular.ttf")
 
@@ -42,6 +55,13 @@ divider={
   layout_marginRight="4dp";
 }
 
+--当系统不支持时占位
+placeholderView={
+  View;
+  layout_width=0,
+  layout_height=0,
+}
+
 ---运行代码，并pcall，防止代码报错
 ---@param code string
 ---@param code showError
@@ -50,8 +70,17 @@ function pcallLoadCode(code,showError)
   local success,value=pcall(assert(load("return "..code)))
   if success then
     return value
-   elseif showError then
-    return "您的设备或编译器无此资源资源:\n"..value
+   else
+    --只有编辑器问题才允许继续执行
+    local isEditorError=value:match("is not a field or mothod")
+    or value:match("Class not found: .-%.R")
+    or value:match("attempt to index a nil value %(global 'androidhwext'%)")
+    if not isEditorError then
+      error(value,2)
+    end
+    if showError then
+      return "您的设备或编译器无此资源资源:\n"..value
+    end
   end
 end
 
@@ -99,6 +128,34 @@ function buildColorShow(zhName,enName,callName,enableSwitchView)
   };
 end
 
+---生成背景展示布局
+---@param zhName string
+---@param enName string
+---@param callName string
+---@param enableSwitchView boolean
+---@return table layout 生成的布局
+function buildBgShow(zhName,enName,callName,enableSwitchView)
+  local title=zhName.."\n"..enName
+  local text=title.."\n  "..callName
+  local spannableText=SpannableString(text)
+
+  return {
+    HorizontalScrollView;
+    layout_width="fill";
+    fillViewport="true";
+    {
+      enableSwitchView and Switch or TextView,
+      text=spannableText,
+      background=pcallLoadCode(callName,false);
+      layout_width="fill",
+      padding="8dp";
+      paddingTop="4dp";
+      paddingBottom="4dp";
+      checked=enableSwitchView and "true";
+      textIsSelectable="true";
+    }
+  };
+end
 ---生成值展示布局
 ---@param resIdCallName string
 ---@param callNames string[]
@@ -154,6 +211,48 @@ local afterLoadLayTime=System.currentTimeMillis()
 table.insert(layout[2],{
   TextView,
   text="获取资源耗时:"..afterLoadLayTime-afterLoadModuleTime,
+  textColor=android.res.color.attr.textColorPrimary;
+  layout_width="fill",
+  padding="4dp";
+})
+
+for i=1,100000 do
+  local colorAccent=android.res.color.attr.colorAccent
+end
+local after100000GetColorTestTime=System.currentTimeMillis()
+
+table.insert(layout[2],{
+  TextView,
+  text="获取统100000次颜色耗时:"..after100000GetColorTestTime-afterLoadLayTime,
+  textColor=android.res.color.attr.textColorPrimary;
+  layout_width="fill",
+  padding="4dp";
+})
+
+for i=1,1000 do
+  local sym_def_app_icon=android.res.drawable.sym_def_app_icon
+end
+local after1000GetDrawableTestTime=System.currentTimeMillis()
+
+table.insert(layout[2],{
+  TextView,
+  text="获取统1000次图片耗时:"..after1000GetDrawableTestTime-after100000GetColorTestTime,
+  textColor=android.res.color.attr.textColorPrimary;
+  layout_width="fill",
+  padding="4dp";
+})
+
+--设置缓存图片，请谨慎使用，因为这种方式可能导致状态，还会造成内存泄露
+res.setSupportCacheGetterMap("getDrawable",true)
+
+for i=1,100000 do
+  local sym_def_app_icon=android.res.drawable.sym_def_app_icon
+end
+local after100000GetCacheDrawableTestTime=System.currentTimeMillis()
+
+table.insert(layout[2],{
+  TextView,
+  text="获取统100000次缓存后图片耗时:"..after100000GetCacheDrawableTestTime-after1000GetDrawableTestTime,
   textColor=android.res.color.attr.textColorPrimary;
   layout_width="fill",
   padding="4dp";